LPS-118825 Move form components into formComponents
